requirements.txt
pyttsx3
speechrecognition
pyaudio
pyautogui
fuzzywuzzy
openai
tk
tkvideo
requests
webbrowser
pywhatkit
python-Levenshtein
pystray
Pillow
